[feature-common] move common class
1. move Post data class from post module to common module
2. move SingleEvent class from post module to common module
