+++ features/post/src/main/java/com/antique/post/view/PostFragment.kt
@@ -3,13 +3,11 @@ package com.antique.post.view
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextWatcher
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.MenuItem
 import android.view.View
 import android.view.ViewGroup
-import android.view.WindowManager
 import androidx.core.view.ViewCompat
 import androidx.core.view.updatePadding
 import androidx.databinding.DataBindingUtil
@@ -17,7 +15,7 @@ import androidx.navigation.fragment.findNavController
 import androidx.navigation.navGraphViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.antique.common.ApiStatus
-import com.antique.post.EventObserver
+import com.antique.common.EventObserver
 import com.antique.post.R
 import com.antique.post.adapter.SelectedImageListAdapter
 import com.antique.post.databinding.FragmentPostBinding
